Adding source code for Spline generation, WormSkeleton representation and WormDescriptor.
WormDescriptor relates a WormSkeleton and the EVImage representation, containing
associated functionality
WormSkeleton defines a worm as a single structure. The idea is making it different from
WormClusterSkeleton to provide specific functionality and gain expressivity

The spline is generated using Curve API which is added as jar file. A Cardinal Spline is generated
to obtain a curve representation that passes through a given set of points. Is represented
in terms of Endrov in EVCardinalSpline, to make renderable on Endrov image data